requirements-dev.txt
# 测试依赖
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
Pillow>=10.0.0
